Rename caching backends in order to not cause any name collisions
All caching backends are now named using a consistent naming scheme:
"<backend type>cache.py".

Fixes #162
