+++ flask_caching/backends/__init__.py
@@ -9,17 +9,17 @@
     :copyright: (c) 2010 by Thadeus Burgess.
     :license: BSD, see LICENSE for more details.
 """
-from flask_caching.backends.filesystem import FileSystemCache
+from flask_caching.backends.filesystemcache import FileSystemCache
 from flask_caching.backends.memcache import (
     MemcachedCache,
     SASLMemcachedCache,
     SpreadSASLMemcachedCache,
 )
-from flask_caching.backends.null import NullCache
+from flask_caching.backends.nullcache import NullCache
 
 # TODO: Rename to "redis" when python2 support is removed
 from flask_caching.backends.rediscache import RedisCache, RedisSentinelCache
-from flask_caching.backends.simple import SimpleCache
+from flask_caching.backends.simplecache import SimpleCache
 
 try:
     from flask_caching.backends.uwsgicache import UWSGICache
